Update and rename settings.js to settings1.js
𝑗𝑎𝑘𝑠𝑜ꪀ 𝑣𝟸𝟼 𔒝
